requirements.txt
fastmcp
requests
pandas
ta
google-generativeai
google-adk
praw
matplotlib
plotly
